+++ CHANGELOG.md
@@ -3,11 +3,14 @@ CHANGELOG
 
 [//]: # (For more details refer to the [documentation]&#40;&#41;.)
 
-Upcoming
+2023-11-02 (0.3.0)
 --------
 * added `json` and `pickle` output file support in `execution.py`
 * added `examples`, `tests` and `show-cov` commands for `poetry`
 * added strict type checking with `mypy` to help catch type issues
+* added commandline function for analysis script in `execution.py`
+* improved CI/CD and added integration tests
+* added check for `cost_function` property in `pre.py`
 
 
 2023-10-23 (0.2.0)